services/backend/requirements.txt
# Coroutines
celery==5.2.7

# Database
psycopg2-binary==2.9.3
redis==4.3.4

# Utils
flake8==4.0.1
flower==1.0.0
gunicorn==20.1.0
isort==5.10.1
python-decouple==3.6
uwsgi==2.0.20

# Websockets
channels==3.0.5
channels-redis==3.4.1
daphne==3.0.2

# Django
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-extensions==3.2.0
django-feature-policy==3.1.0
django-filter==22.1
django-htmlmin==0.11.0
django-ipware==4.0.2
django-oauth-toolkit==2.1.0
django-otp==1.1.3
django-postgres-extensions-up==0.9.4
django-redis==5.2.0
django-referrer-policy==1.0
django-rest-swagger==2.2.0
django-simple-history==3.1.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0

# Media
boto3==1.24.79
whitenoise==6.2.0

# Testing
pytest==6.2.5

# Logging
sentry-sdk==1.9.3
